#4 Fix failing compilation error
Target need to be hooked befre CollectPackageReferences so the Restore target has already the data. This target is unfortunately run twice. Subsequent target CollectPackageDownloads is run once but too late for the custom tasks' changes to be noticed.

The problem could be probably resolved by hooking the tasks to one of the internal targets that are run during CollectPackageReferences, such as `_GenerateRestoreSpecs` or `_GetRestoreSettings`. This would require some time and trial and fail method. Would be resolved in other issue.